requirements.txt
requests
Flask
python_dotenv==1.0.1